Added Edit link for data release updates if you are admin. Also moved getUserGroups fxn to utils/JWT from submission-view
